refactor: 상품 상세 bottomBar 컴포넌트 분리
